Collision Handler expanded for support of detailed collision detection (binary search), trigger system is not finished.

+++ app/src/main/java/com/demoncube/ninjaadventure/game/controlers/ControllerInterface.java
@@ -7,6 +7,6 @@ public interface ControllerInterface {
     Victor getMoveVectors();
     boolean isAttacking();
     boolean isInteracting();
-    void feedBack( int[] feedBack);
+    void feedBack( double[] feedBack);
 
 }
